requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
alembic==1.14.1
async-timeout==5.0.1
asyncio-redis==0.16.0
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
ecdsa==0.19.0
greenlet==3.1.1
html5lib==1.1
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
redis==5.2.1
redis-cli==1.0.1
requests==2.32.3
rsa==4.9
ruff==0.9.4
sanic==24.12.0
Sanic-Auth==0.3.0
sanic-ext==23.12.0
sanic-jinja2==2022.11.11
sanic-jwt==1.8.0
sanic-routing==23.12.0
sanic-session==0.8.0
six==1.17.0
SQLAlchemy==2.0.37
tracerite==1.1.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
